server/requirements.txt
alembic==1.8.0
bcrypt==3.2.2
email-validator==1.2.1
fastapi==0.78.0
passlib==1.7.4
psycopg2==2.9.3
pydantic==1.9.1
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
uvicorn==0.17.6
